add uuidToColor to utils file
